correct lldb path. replace tab with space.
With a certain version of lldb and continuing the use of `sys.executable`,
installation will go wrong. The lldb version of v360.99.0 will output
differently than the others.
In further detail, the other lldb's will output the path to the python executable

    /usr/local/opt/python/bin/python2.7

while the version I had will output the path to the lldb executable.

    /usr/local/opt/llvm/bin/lldb

With this in mind, this commit has a workaround of retrieving the python
executable for lldb without problems.